// packages/pentatrion-geo/src/geocode/ign-geocodage-api.d.ts
export type APISchemas = {
// ...
  AddressProperties: {
    /* Libellé complet de l'adresse */
    label?: string;
    id?: string;
    postcode?: APISchemas["PostalCode"];
    /* Commune de l'adresse */
    city?: string;
    /* Arrondissement de l'adresse */
    disctrict?: string;
    /* Rue de l'adresse */
    street?: string;
    housenumber?: string;
    citycode?: APISchemas["InseeCode"];
    /* Longitude de l'adresse */
    x?: number;
    /* Latitude de l'adresse */
    y?: number;
    score?: number;
    _score?: number;
    name?: string;
    type?: "housenumber" | "street" | "locality" | "municipality";
    /* Rétro-compatibilité */
    _type?: "address";
    context?: string;
    importance?: number;
  };
// ...
};